no more need to call ->value
The reach constructor set error_handler and fault_handler to die
so the errors can ben caught by eval.

This is made because i wanted the metaprogrammed methods to call
->value so the library is easier to use.

I also added notes about a singleton lib. it would be much handy to use
for tasks requiring a single wiki (99% of the time?).